biometric_auth/requirements.txt
Flask==3.0.0
onnxruntime==1.17.0
opencv-python==4.9.0.80
numpy==1.26.3
cryptography==42.0.2
scipy==1.12.0
Pillow==10.2.0
mediapipe==0.10.21
requests==2.31.0
